requirements.txt
networkx>=3.0
matplotlib>=3.7
pyyaml>=6.0
numpy>=1.24
